Struct mbox::mbox::MBox

pub struct MBox<T: ?Sized + Free>(_);

A malloc-backed box. This structure allows Rust to exchange objects with C without cloning.

Methods

impl<T: ?Sized + Free> MBox<T>

pub unsafe fn from_raw(ptr: *mut T) -> MBox<T>

Constructs a new malloc-backed box from a pointer allocated by malloc. The content of the pointer must be already initialized.

pub fn as_ptr(&self) -> *const T

Obtains the pointer owned by the box.

pub fn as_mut_ptr(&mut self) -> *mut T

Obtains the mutable pointer owned by the box.

impl<T: ?Sized + Free> MBox<T>

pub fn into_raw(self) -> *mut T

Consumes the box and returns the original pointer.

The caller is responsible for freeing the pointer after this.

impl<T> MBox<T>

pub fn new(value: T) -> MBox<T>

Constructs a new malloc-backed box, and move an initialized value into it.

impl<T> MBox<[T]>

pub unsafe fn from_raw_parts(value: *mut T, len: usize) -> MBox<[T]>

Constructs a new malloc-backed slice from the pointer and the length (number of items).

The malloced size of the pointer must be at least len * size_of::<T>(). The content must already been initialized.

impl<T: Clone> MBox<[T]>

pub fn from_slice(slice: &[T]) -> MBox<[T]>

Creates a new malloc-boxed slice by cloning the content of an existing slice.

impl MBox<str>

pub unsafe fn from_raw_utf8_parts_unchecked(
value: *mut u8,
len: usize
) -> MBox<str>

Constructs a new malloc-backed string from the pointer and the length (number of UTF-8 code units).

The malloced size of the pointer must be at least len. The content must already been initialized and be valid UTF-8.

pub unsafe fn from_raw_utf8_parts(
value: *mut u8,
len: usize
) -> Result<MBox<str>, Utf8Error>

Constructs a new malloc-backed string from the pointer and the length (number of UTF-8 code units).

The malloced size of the pointer must be at least len. If the content does not contain valid UTF-8, this method returns an Err.

pub fn into_bytes(self) -> MBox<[u8]>

Converts the string into raw bytes.

pub unsafe fn from_utf8_unchecked(bytes: MBox<[u8]>) -> MBox<str>

Creates a string from raw bytes. The bytes must be valid UTF-8.

pub fn from_utf8(bytes: MBox<[u8]>) -> Result<MBox<str>, Utf8Error>

Creates a string from raw bytes. If the content does not contain valid UTF-8, this method returns an Err.

pub fn from_str(string: &str) -> MBox<str>

Creates a new malloc-boxed string by cloning the content of an existing string slice.